chore(toolkit-lib): unify IoHost helper
Previously we had two different IoHost helpers for use in tests with slightly different functionality. I don't see a reason why this needs to be an either or, so I combined them into one.

Also gets rid of RequireApproval check in the TestIoHost and a corresponding test case. This test case was testing the implementation of the TestIoHost only, which we obviously do not care about. With that change, we can also remove a leftover enum.
