+++ packages/@aws-cdk/toolkit-lib/lib/api/index.ts
@@ -13,7 +13,6 @@ export * from './logs-monitor';
 export * from './notices';
 export * from './plugin';
 export * from './refactoring';
-export * from './require-approval';
 export * from './resource-import';
 export * from './rwlock';
 export * from './settings';
