--- packages/@aws-cdk/toolkit-lib/lib/api/require-approval.ts
@@ -1,17 +1,0 @@
-/**
- * @deprecated
- */
-export enum RequireApproval {
-  /**
-   * Never require any security approvals
-   */
-  NEVER = 'never',
-  /**
-   * Any security changes require an approval
-   */
-  ANY_CHANGE = 'any-change',
-  /**
-   * Require approval only for changes that are access broadening
-   */
-  BROADENING = 'broadening',
-}
