+++ packages/@aws-cdk/toolkit-lib/test/api/environment/environment-resources.test.ts
@@ -5,7 +5,6 @@ import { EnvironmentResourcesRegistry } from '../../../lib/api/environment';
 import { Notices } from '../../../lib/api/notices';
 import { CachedDataSource } from '../../../lib/api/notices/cached-data-source';
 import { NoticesFilter } from '../../../lib/api/notices/filter';
-import { FakeIoHost } from '../../_helpers/fake-io-host';
 import { MockSdk, mockBootstrapStack, mockSSMClient } from '../../_helpers/mock-sdk';
 import { TestIoHost } from '../../_helpers/test-io-host';
 import { MockToolkitInfo } from '../_helpers/mock-toolkitinfo';
@@ -103,7 +102,7 @@ describe('validateversion without bootstrap stack', () => {
       .mockImplementation(() => Promise.resolve({ expiration: 0, notices: [] }));
 
     // THEN
-    const myIoHost = new FakeIoHost();
+    const myIoHost = new TestIoHost();
     const notices = Notices.create({ context: new Context(), ioHost: myIoHost, cliVersion: '1.0.0' });
     await notices.refresh({ dataSource: { fetch: async () => [] } });
     await expect(envResources().validateVersion(8, '/abc')).resolves.toBeUndefined();
